# backend/internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/account": {
            "get": {
                "tags": [
                    "account"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page_index",
                        "name": "page_index",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page_size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name or account",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "account id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "account ids",
                        "name": "ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "account name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "is info mode",
                        "name": "info",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "account type",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.HttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/controller.ListData"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/model.Account"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "account"
                ],
                "parameters": [
                    {
                        "description": "account",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Account"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.HttpResponse"
                        }
                    }
                }
            }
        },
        "/account/:id": {
            "put": {
                "tags": [
                    "account"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "account id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "account",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Account"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.HttpResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "account"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "account id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.HttpResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.HttpResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "controller.ListData": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "list": {
                    "type": "array",
                    "items": {}
                }
            }
        },
...
        "model.Account": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "account_type": {
                    "type": "integer"
                },
                "asset_count": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "phrase": {
                    "type": "string"
                },
                "pk": {
                    "type": "string"
                },
                "resource_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "updater_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}